Refactor app settings initialization codes
